@@ -1,6 +1,6 @@
 {
-  "name": "justcoded/yii2-rbac",
-  "description": "Yii2 RBAC Module",
+  "name": "deadmantfa/yii2-rbac",
+  "description": "Yii2 RBAC Module, updated for PHP 8 and above",
   "type": "yii2-extension",
   "keywords": [
     "yii2",
@@ -13,17 +13,33 @@
     {
       "name": "Alex Prokopenko",
       "email": "aprokopenko@justcoded.com"
+    },
+    {
+      "name": "Wenceslaus Dsilva",
+      "email": "wenceslausdsilva@gmail.com"
     }
   ],
-  "minimum-stability": "dev",
+  "minimum-stability": "stable",
   "require": {
-    "php": ">=7.0.0",
-    "yiisoft/yii2": "^2.0.12",
-    "kartik-v/yii2-widget-select2": "^2.1.0"
+    "php": ">=8.0.0",
+    "yiisoft/yii2": "^2.0.51",
+    "kartik-v/yii2-widget-select2": "^2.2.0",
+    "yiisoft/yii2-bootstrap4": "^2.0"
   },
   "autoload": {
     "psr-4": {
-      "justcoded\\yii2\\rbac\\": "src/"
+      "deadmantfa\\yii2\\rbac\\": "src/"
+    }
+  },
+  "repositories": [
+    {
+      "type": "composer",
+      "url": "https://asset-packagist.org"
+    }
+  ],
+  "config": {
+    "allow-plugins": {
+      "yiisoft/yii2-composer": true
     }
   }
 }
