@@ -1,304 +1,234 @@
 <?php
 
-namespace justcoded\yii2\rbac\commands;
+declare(strict_types=1);
 
-use justcoded\yii2\rbac\forms\ScanForm;
-use justcoded\yii2\rbac\models\Item;
+namespace deadmantfa\yii2\rbac\commands;
+
+use deadmantfa\yii2\rbac\forms\ScanForm;
+use deadmantfa\yii2\rbac\models\Item;
 use Yii;
-use yii\base\InvalidParamException;
+use yii\base\Exception;
 use yii\console\Controller;
 use yii\console\ExitCode;
 use yii\helpers\Console;
-use yii\helpers\Inflector;
 use yii\rbac\Permission;
 use yii\rbac\Role;
 use yii\rbac\Rule;
 
 class RbacController extends Controller
 {
-	/**
-	 * @var string Path to scan.
-	 */
-	public $path = '@app';
+    private const PROMPT_CONFIRMATION = 'Are you sure you want to continue? (yes/no)';
 
-	/**
-	 * @var string Paths to ignore.
-	 * Use comma to specify several paths.
-	 */
-	public $ignorePath = [];
+    /**
+     * @var string Path to scan.
+     */
+    public string $path = '@app';
 
-	/**
-	 * @var string Routes base prefix to be added to all found routes
-	 */
-	public $routesBase = '';
+    /**
+     * @var array Paths to ignore.
+     */
+    public array $ignorePath = [];
 
-	/**
-	 * @inheritdoc
-	 *
-	 * @return array|string[]
-	 */
-	public function options($actionID)
-	{
-		if ('scan' === $actionID) {
-			return ['path', 'ignorePath', 'routesBase'];
-		} else {
-			return parent::options($actionID);
-		}
-	}
+    /**
+     * @inheritdoc
+     */
+    public function options($actionID): array
+    {
+        return $actionID === 'scan' ? ['path', 'ignorePath'] : parent::options($actionID);
+    }
 
-	/**
-	 * @inheritdoc
-	 *
-	 * @return array
-	 */
-	public function optionAliases()
-	{
-		return [
-			'p' => 'path',
-			'i' => 'ignorePath',
-			'b' => 'routesBase',
-		];
-	}
+    /**
+     * @inheritdoc
+     */
+    public function optionAliases(): array
+    {
+        return [
+            'p' => 'path',
+            'i' => 'ignorePath',
+        ];
+    }
 
-	/**
-	 * Init default roles, master and administer permissions.
-	 *
-	 * @return int
-	 */
-	public function actionInit()
-	{
-		$this->warning('Init command will clean all your current roles, permissions and assignments.');
-		$confirm = $this->prompt('Are you sure you want to continue? (yes/no)', [
-			'required'  => true,
-		]);
-		if ('y' !== strtolower($confirm){0}) {
-			$this->line("\nTerminating.");
-			return ExitCode::DATAERR;
-		}
+    /**
+     * Initializes default roles, master, and administer permissions.
+     *
+     * @return int
+     * @throws Exception
+     */
+    public function actionInit(): int
+    {
+        $this->log('Init command will clean all your current roles, permissions, and assignments.', Console::FG_YELLOW);
+        $confirm = $this->prompt(self::PROMPT_CONFIRMATION, ['required' => true]);
+        if (strtolower($confirm)[0] !== 'y') {
+            $this->log("\nTerminating.", Console::FG_RED);
+            return ExitCode::DATAERR;
+        }
 
-		$am = Yii::$app->authManager;
-		$am->removeAll();
+        $am = Yii::$app->authManager;
+        $am->removeAll();
 
-		$everything = $this->addPermission(Item::PERMISSION_MASTER, 'Allow Everything');
-		$administer = $this->addPermission(Item::PERMISSION_ADMINISTER, 'Access administration panel.');
+        $everything = $this->addPermission(Item::PERMISSION_MASTER, 'Allow Everything');
+        $administer = $this->addPermission(Item::PERMISSION_ADMINISTER, 'Access administration panel.');
 
-		$guest = $this->addRole(Item::ROLE_GUEST, 'Usual site visitor.');
+        $this->addRole(Item::ROLE_GUEST, 'Usual site visitor.');
+        $this->addRole(Item::ROLE_AUTHENTICATED, 'Authenticated user.');
+        $this->addRole(Item::ROLE_ADMIN, 'Administrator.', [$administer]);
+        $this->addRole(Item::ROLE_MASTER, 'Has full system access.', [$everything]);
 
-		$authenticated = $this->addRole(Item::ROLE_AUTHENTICATED, 'Authenticated user.');
+        $this->log("Finished.", Console::FG_GREEN);
+        return ExitCode::OK;
+    }
 
-		$admin = $this->addRole(Item::ROLE_ADMIN, 'Administrator.', [
-			$administer
-		]);
+    /**
+     * Unified logging method with ANSI color support.
+     */
+    protected function log(string $message, int $color = Console::FG_GREY): void
+    {
+        $this->stdout($message . "\n", $color);
+    }
 
-		$sysAdmin = $this->addRole(Item::ROLE_MASTER, 'Has full system access.', [
-			$everything,
-		]);
+    /**
+     * Adds a permission to the RBAC system.
+     *
+     * @param string $name
+     * @param string $descr
+     * @param Rule|null $rule
+     * @param array $parents
+     * @return Permission
+     * @throws Exception
+     * @throws \Exception
+     */
+    protected function addPermission(string $name, string $descr, Rule $rule = null, array $parents = []): Permission
+    {
+        $auth = Yii::$app->authManager;
 
-		$this->success("Finished.");
-		return ExitCode::OK;
-	}
+        $permission = $auth->createPermission($name);
+        $permission->description = $descr;
+        if ($rule) {
+            $permission->ruleName = $rule->name;
+        }
+        $auth->add($permission);
+        $this->log("Added permission: $permission->name", Console::FG_GREEN);
 
-	/**
-	 * Assign master role with full system access to a user.
-	 *
-	 * @param integer $userId
-	 *
-	 * @return int
-	 */
-	public function actionAssignMaster($userId)
-	{
-		if (!is_numeric($userId) || 0 >= $userId) {
-			$this->error("User ID should be a positive integer");
-			return ExitCode::DATAERR;
-		}
+        foreach ($parents as $parent) {
+            $auth->addChild($parent, $permission);
+            $this->log("Added as child of: $parent->name", Console::FG_GREEN);
+        }
 
-		$am = Yii::$app->authManager;
-		if (!$master = $am->getRole(Item::ROLE_MASTER)) {
-			$this->error("Master role not found. Please run 'init' command first.");
-			return ExitCode::UNAVAILABLE;
-		}
+        return $permission;
+    }
 
-		$am->assign($master, $userId);
-		$this->success("Role ".Item::ROLE_MASTER." added to user {$userId}.");
+    /**
+     * Adds a role to the RBAC system.
+     *
+     * @param string $name
+     * @param string $descr
+     * @param array $childs
+     * @return Role
+     * @throws Exception
+     * @throws \Exception
+     */
+    protected function addRole(string $name, string $descr, array $childs = []): Role
+    {
+        $auth = Yii::$app->authManager;
 
-		return ExitCode::OK;
-	}
+        $role = $auth->createRole($name);
+        $role->description = $descr;
+        $auth->add($role);
+        $this->log("Added role: $role->name", Console::FG_GREEN);
 
-	/**
-	 * Assign permissions or roles to users
-	 * You can pass several permissions or users using comma.
-	 *
-	 * @param array $items
-	 * @param array $userIds
-	 *
-	 * @return int
-	 */
-	public function actionAssign(array $items, array $userIds)
-	{
-		$this->line('Assigning items...');
-		$am = Yii::$app->authManager;
+        foreach ($childs as $child) {
+            $auth->addChild($role, $child);
+            $this->log("Added '$child->name' to $role->name", Console::FG_GREEN);
+        }
 
-		foreach ($items as $item) {
-			$this->warning('Trying to assign ' . $item);
-			// check role exists
-			if (! $authItem = $am->getRole($item)) {
-				if (! $authItem = $am->getPermission($item)) {
-					$this->error("\trole or permission not found");
-					continue;
-				}
-			}
+        return $role;
+    }
 
-			// assign
-			foreach ($userIds as $userId) {
-				$am->assign($authItem, $userId);
-				$this->line("\tassigned to user " . $userId);
-			}
-		}
+    /**
+     * Assigns the master role to a user.
+     *
+     * @param int $userId
+     * @return int
+     * @throws \Exception
+     */
+    public function actionAssignMaster(int $userId): int
+    {
+        if ($userId <= 0) {
+            $this->log("User ID should be a positive integer.", Console::FG_RED);
+            return ExitCode::DATAERR;
+        }
 
-		$this->success("Finished.");
-		return ExitCode::OK;
-	}
+        $am = Yii::$app->authManager;
+        $master = $am->getRole(Item::ROLE_MASTER);
+        if (!$master) {
+            $this->log("Master role not found. Please run 'init' command first.", Console::FG_RED);
+            return ExitCode::UNAVAILABLE;
+        }
 
-	/**
-	 * Scans controller/module directories to get route permissions info
-	 *
-	 *      runs recursive search for *Controller.php, ignores files, which specified in ignorePath
-	 *      register auth Permissions based on Controller public actions (methods in format "action{something}")
-	 *
-	 * @return int
-	 */
-	public function actionScan()
-	{
-		$scanner = new ScanForm();
-		$scanner->path = $this->path;
-		$scanner->ignorePath = $this->ignorePath;
-		$scanner->routesBase = $this->routesBase;
+        $am->assign($master, $userId);
+        $this->log("Role " . Item::ROLE_MASTER . " added to user $userId.", Console::FG_GREEN);
 
-		$this->warning('Scanning directory for controllers and routes...');
+        return ExitCode::OK;
+    }
 
-		if ($actionRoutes = $scanner->scan()) {
-			$this->line("\tfound " . count($actionRoutes) . ' routes');
-			$inserted = $scanner->importPermissions($actionRoutes);
-			foreach ($inserted as $route => $status) {
-				$this->line("\t\tadded " . $route);
-			}
+    /**
+     * Assigns permissions or roles to users.
+     *
+     * @param array $items
+     * @param array $userIds
+     * @return int
+     * @throws \Exception
+     */
+    public function actionAssign(array $items, array $userIds): int
+    {
+        if (empty($items) || empty($userIds)) {
+            $this->log("Both items and user IDs are required.", Console::FG_RED);
+            return ExitCode::DATAERR;
+        }
 
-			$inserted = count($inserted);
-			$this->success("Inserted {$inserted} new permissions.");
-		} else {
-			$this->line("No routes to import");
-		}
+        $am = Yii::$app->authManager;
 
-		return ExitCode::OK;
-	}
+        foreach ($items as $item) {
+            $authItem = $am->getRole($item) ?? $am->getPermission($item);
+            if (!$authItem) {
+                $this->log("Role or permission '$item' not found.", Console::FG_RED);
+                continue;
+            }
 
-	/**
-	 * @param string    $name
-	 * @param string    $descr
-	 * @param Rule|null $rule
-	 * @param Permission[]|Role[]     $parents  assign permission to some parent
-	 *
-	 * @return Permission
-	 */
-	protected function addPermission($name, $descr, $rule = null, $parents = [])
-	{
-		$auth = Yii::$app->authManager;
+            foreach ($userIds as $userId) {
+                Yii::$app->authManager->assign($authItem, $userId);
+                $this->log("Assigned '$authItem->name' to user $userId.", Console::FG_GREEN);
+            }
+        }
 
-		$p = $auth->createPermission($name);
-		$p->description = $descr;
-		if ($rule) {
-			$p->ruleName = $rule->name;
-		}
-		$auth->add($p);
-		$this->line("\tadded permission " . $p->name);
+        $this->log("Finished assigning items.", Console::FG_GREEN);
+        return ExitCode::OK;
+    }
 
-		foreach ($parents as $parent) {
-			$auth->addChild($parent, $p);
-			$this->line("\t\tadded as child of " . $parent->name);
-		}
+    /**
+     * Scans controller/module directories to get route permissions info.
+     *
+     * @return int
+     * @throws Exception
+     */
+    public function actionScan(): int
+    {
+        $scanner = new ScanForm();
+        $scanner->path = $this->path;
+        $scanner->ignorePath = $this->ignorePath;
 
-		return $p;
-	}
+        $this->log('Scanning directory for controllers and routes...', Console::FG_YELLOW);
 
-	/**
-	 * @param string              $name
-	 * @param string              $descr
-	 * @param Permission[]|Role[] $childs
-	 *
-	 * @return Role
-	 */
-	protected function addRole($name, $descr, $childs = [])
-	{
-		$auth = Yii::$app->authManager;
+        $actionRoutes = $scanner->scan();
+        if ($actionRoutes) {
+            $this->log("Found " . count($actionRoutes) . ' routes.', Console::FG_GREEN);
+            foreach ($scanner->importPermissions($actionRoutes) as $route => $status) {
+                $this->log("Added route: $route", Console::FG_GREEN);
+            }
+        } else {
+            $this->log("No routes found during the scan.", Console::FG_YELLOW);
+        }
 
-		$r = $auth->createRole($name);
-		$r->description = $descr;
-		$auth->add($r);
-		$this->warning("Added {$r->name} role");
-
-		if ($childs) {
-			foreach ($childs as $child) {
-				$auth->addChild($r, $child);
-				$this->line("\tAdded '{$child->name}' to {$r->name}");
-			}
-		}
-
-		return $r;
-	}
-
-	/**
-	 * Prints a line to STDOUT.
-	 *
-	 * You may optionally format the string with ANSI codes by
-	 * passing additional parameters using the constants defined in [[\yii\helpers\Console]].
-	 *
-	 * Example:
-	 *
-	 * ```
-	 * $this->stdout('This will be red and underlined.', Console::FG_RED, Console::UNDERLINE);
-	 * ```
-	 *
-	 * @param string $string the string to print
-	 * @return int|bool Number of bytes printed or false on error
-	 */
-	protected function line($string = '')
-	{
-		$args = func_get_args();
-		$args[0] .= "\n";
-		return call_user_func_array([$this, 'stdout'], $args);
-	}
-
-	/**
-	 * Prints a yellow line to STDOUT.
-	 *
-	 * @param string $string the string to print
-	 * @return int|bool Number of bytes printed or false on error
-	 */
-	protected function warning($string)
-	{
-		return $this->line($string, Console::FG_YELLOW);
-	}
-
-	/**
-	 * Prints a yellow line to STDOUT.
-	 *
-	 * @param string $string the string to print
-	 * @return int|bool Number of bytes printed or false on error
-	 */
-	protected function success($string)
-	{
-		return $this->line($string, Console::FG_GREEN);
-	}
-
-	/**
-	 * Prints a red line to STDOUT.
-	 *
-	 * @param string $string the string to print
-	 * @return int|bool Number of bytes printed or false on error
-	 */
-	protected function error($string)
-	{
-		return $this->line($string, Console::FG_RED);
-	}
-
+        return ExitCode::OK;
+    }
 }
