@@ -1,69 +1,69 @@
 <?php
 
-namespace justcoded\yii2\rbac\forms;
+declare(strict_types=1);
+
+namespace deadmantfa\yii2\rbac\forms;
 
 use yii\base\Model;
-use Yii;
-use yii\helpers\ArrayHelper;
 
 abstract class ItemForm extends Model
 {
-	const SCENARIO_CREATE = 'create';
+    const SCENARIO_CREATE = 'create';
 
-	public $name;
-	public $type;
-	public $description;
-	public $ruleName;
-	public $data;
-	public $createdAt;
-	public $updatedAt;
+    public string $name;
+    public int $type;
+    public ?string $description;
+    public ?string $ruleName;
+    public ?string $data;
+    public ?int $createdAt;
+    public ?int $updatedAt;
 
-	/**
-	 * @return array
-	 */
-	public function rules()
-	{
-		return  [
-			['name', 'match', 'pattern' => static::getNamePattern()],
-			[['type', 'name'], 'required'],
-			['name', 'uniqueItemName', 'on' => static::SCENARIO_CREATE],
-			[['name', 'ruleName'], 'trim'],
-			[['name', 'description', 'ruleName', 'data'], 'string'],
-			[['type', 'createdAt', 'updatedAt'], 'integer'],
-		];
-	}
+    /**
+     * @return array
+     */
+    public function rules(): array
+    {
+        return [
+            ['name', 'match', 'pattern' => static::getNamePattern()],
+            [['type', 'name'], 'required'],
+            ['name', 'uniqueItemName', 'on' => static::SCENARIO_CREATE],
+            [['name', 'ruleName'], 'trim'],
+            [['name', 'description', 'ruleName', 'data'], 'string'],
+            [['type', 'createdAt', 'updatedAt'], 'integer'],
+        ];
+    }
 
-	/**
-	 * Validate item (permission/role) name to be unique
-	 *
-	 * @param string $attribute
-	 * @param array  $params
-	 * @param mixed  $validator
-	 *
-	 * @return bool
-	 */
-	abstract public function uniqueItemName($attribute, $params, $validator);
+    /**
+     * RBAC Item name validation pattern
+     *
+     * @return string
+     */
+    public static function getNamePattern(): string
+    {
+        return '/^[a-z0-9\s\_\-\/\*]+$/i';
+    }
 
-	/**
-	 * @inheritdoc
-	 */
-	public function attributeLabels()
-	{
-		return [
-			'name' => 'Name',
-			'description' => 'Description',
-			'ruleName' => 'Rule Class',
-		];
-	}
+    /**
+     * Validate item (permission/role) name to be unique
+     *
+     * @param string $attribute
+     * @param array $params
+     * @param mixed $validator
+     *
+     * @return bool
+     */
+    abstract public function uniqueItemName(string $attribute, array $params, mixed $validator): bool;
 
-	/**
-	 * RBAC Item name validation pattern
-	 *
-	 * @return string
-	 */
-	public static function getNamePattern()
-	{
-		return '/^[a-z0-9\s\_\-\/\*]+$/i';
-	}
+    /**
+     * @inheritdoc
+     */
+    public function attributeLabels(): array
+    {
+        return [
+            'name' => 'Name',
+            'description' => 'Description',
+            'ruleName' => 'Rule Class',
+        ];
+    }
 
 }
