@@ -1,122 +1,133 @@
 <?php
 
-namespace justcoded\yii2\rbac\models;
+declare(strict_types=1);
 
+namespace deadmantfa\yii2\rbac\models;
+
+use Yii;
 use yii\base\Model;
 use yii\data\ArrayDataProvider;
-use yii\rbac\Role as RbacRole;
 use yii\rbac\Permission as RbacPermission;
-use Yii;
+use yii\rbac\Role as RbacRole;
 
 
 class ItemSearch extends Model
 {
-	public $roleName;
-	public $permName;
-	public $permRole;
+    /**
+     * @var string|null Role name for search.
+     */
+    public ?string $roleName = null;
 
-	/**
-	 * @inheritdoc
-	 */
-	public function rules()
-	{
-		$roles = array_keys(Yii::$app->authManager->getRoles());
-		return [
-			[['roleName', 'permName'], 'string'],
-			[['permRole'], 'in', 'range' => $roles],
-		];
-	}
+    /**
+     * @var string|null Permission name for search.
+     */
+    public ?string $permName = null;
 
-	/**
-	 * @param $params
-	 *
-	 * @return ArrayDataProvider
-	 */
-	public function searchRoles($params)
-	{
-		$this->load($params);
+    /**
+     * @var string|null Selected role name for filtering permissions.
+     */
+    public ?string $permRole = null;
 
-		$roles = Yii::$app->authManager->getRoles();
-		if ($this->roleName) {
-			$roles = array_filter($roles, function (RbacRole $role) {
-				return false !== strpos(strtolower($role->name), $this->roleName);
-			});
-		}
 
-		$dataProvider = new ArrayDataProvider([
-			'allModels'  => $roles,
-			'pagination' => [
-				'pageSize' => 20,
-			],
-		]);
+    /**
+     * @param $permissionName
+     * @return array
+     */
+    public static function getRoleByPermission($permissionName): array
+    {
+        $roles = Yii::$app->authManager->getRoles();
 
-		return $dataProvider;
-	}
+        $array = [];
+        foreach ($roles as $role) {
+            $permissions = Yii::$app->authManager->getPermissionsByRole($role->name);
+            foreach ($permissions as $permission) {
+                if ($permissionName == $permission->name) {
+                    $array[] = $role->name;
+                }
+            }
 
-	/**
-	 * @param $params
-	 *
-	 * @return ArrayDataProvider
-	 */
-	public function searchPermissions($params)
-	{
-		$this->load($params);
+        }
 
-		$permissions = $this->permRole ? Yii::$app->authManager->getPermissionsByRole($this->permRole) :
-			Yii::$app->authManager->getPermissions();
+        return $array;
+    }
 
-		if ($this->permName) {
-			$permissions = array_filter($permissions, function (RbacPermission $permission) {
-				return false !== strpos(strtolower($permission->name), $this->permName);
-			});
-		}
+    /**
+     * @param $parent
+     *
+     * @return array
+     */
+    public static function getInherit($parent): array
+    {
+        $array = [];
+        if ($children = Yii::$app->authManager->getChildren($parent)) {
+            foreach ($children as $child) {
+                if ($child->type == Item::TYPE_ROLE) {
+                    $array[] = $child->name;
+                }
+            }
+        }
+        return $array;
+    }
 
-		$dataProvider = new ArrayDataProvider([
-			'allModels'  => $permissions,
-			'pagination' => [
-				'pageSize' => 20,
-			],
-		]);
+    /**
+     * @inheritdoc
+     */
+    public function rules(): array
+    {
+        $roles = array_keys(Yii::$app->authManager->getRoles());
+        return [
+            [['roleName', 'permName'], 'string'],
+            [['permRole'], 'in', 'range' => $roles],
+        ];
+    }
 
-		return $dataProvider;
-	}
+    /**
+     * @param $params
+     *
+     * @return ArrayDataProvider
+     */
+    public function searchRoles($params): ArrayDataProvider
+    {
+        $this->load($params);
 
-	/**
-	 * @return array
-	 */
-	public static function getRoleByPermission($permissionName)
-	{
-		$roles = Yii::$app->authManager->getRoles();
+        $roles = Yii::$app->authManager->getRoles();
+        if ($this->roleName) {
+            $roles = array_filter($roles, function (RbacRole $role) {
+                return str_contains(strtolower($role->name), $this->roleName);
+            });
+        }
 
-		$array = [];
-		foreach ($roles as $role) {
-			$permissions = Yii::$app->authManager->getPermissionsByRole($role->name);
-			foreach ($permissions as $permission) {
-				if ($permissionName == $permission->name) {
-					$array[] = $role->name;
-				}
-			}
+        return new ArrayDataProvider([
+            'allModels' => $roles,
+            'pagination' => [
+                'pageSize' => 20,
+            ],
+        ]);
+    }
 
-		}
+    /**
+     * @param $params
+     *
+     * @return ArrayDataProvider
+     */
+    public function searchPermissions($params): ArrayDataProvider
+    {
+        $this->load($params);
 
-		return $array;
-	}
+        $permissions = $this->permRole ? Yii::$app->authManager->getPermissionsByRole($this->permRole) :
+            Yii::$app->authManager->getPermissions();
 
-	/**
-	 * @param $parent
-	 *
-	 * @return mixed|string
-	 */
-	public static function getInherit($parent)
-	{
-		$array = [];
-		if ($children = Yii::$app->authManager->getChildren($parent)) {
-			foreach ($children as $child) {
-				if ($child->type == Item::TYPE_ROLE) {
-					$array[] = $child->name;
-				}
-			}
-		}
-		return $array;
-	}
+        if ($this->permName) {
+            $permissions = array_filter($permissions, function (RbacPermission $permission) {
+                return str_contains(strtolower($permission->name), $this->permName);
+            });
+        }
+
+        return new ArrayDataProvider([
+            'allModels' => $permissions,
+            'pagination' => [
+                'pageSize' => 20,
+            ],
+        ]);
+    }
 }
