@@ -1,14 +1,11 @@
 <?php
+declare(strict_types=1);
 
-namespace justcoded\yii2\rbac;
+namespace deadmantfa\yii2\rbac;
 
 
 class Module extends \yii\base\Module
 {
 	public $defaultRoute = 'permissions/index';
 
-	public function init()
-	{
-		parent::init();
-	}
 }
