@@ -1,22 +1,26 @@
 <?php
 
-namespace justcoded\yii2\rbac\assets;
+declare(strict_types=1);
 
-class RbacAssetBundle extends \yii\web\AssetBundle
+namespace deadmantfa\yii2\rbac\assets;
+
+use yii\web\AssetBundle;
+
+class RbacAssetBundle extends AssetBundle
 {
-	public $sourcePath = '@vendor/justcoded/yii2-rbac/src/assets';
+    public $sourcePath = '@vendor/deadmantfa/yii2-rbac/src/assets';
 
-	public $css = [
-		'css/jstree.min.css',
-		'css/custom.css',
-	];
+    public $css = [
+        'css/jstree.min.css',
+        'css/custom.css',
+    ];
 
-	public $js = [
-		'js/jstree.min.js',
-		'js/jstree-double-panel.js',
-		'js/rbac.js'
-	];
-	public $depends = [
-		'yii\web\YiiAsset',
-	];
+    public $js = [
+        'js/jstree.min.js',
+        'js/jstree-double-panel.js',
+        'js/rbac.js'
+    ];
+    public $depends = [
+        'yii\web\YiiAsset',
+    ];
 }
