@@ -1,130 +1,194 @@
 <?php
 
-namespace justcoded\yii2\rbac\helpers;
+declare(strict_types=1);
 
+namespace deadmantfa\yii2\rbac\helpers;
+
+use Exception;
+use RecursiveDirectoryIterator;
+use RecursiveIteratorIterator;
+use RecursiveRegexIterator;
+use ReflectionClass;
+use ReflectionException;
+use ReflectionMethod;
+use RegexIterator;
 use Yii;
 use yii\helpers\Inflector;
+use yii\web\Controller;
 
 class ScanHelper
 {
-	/**
-	 * Recursive scan of directory searching the controllers.
-	 *
-	 * @param string $directory
-	 * @param array  $ignorePath
-	 *
-	 * @return array
-	 */
-	public static function scanControllers(string $directory, array $ignorePath)
-	{
-		// register iterator to find files
-		$dirIt = new \RecursiveDirectoryIterator($directory);
-		$it = new \RecursiveIteratorIterator($dirIt);
+    /**
+     * Recursively scans a directory to find controller files.
+     *
+     * @param string $directory
+     * @param array $ignorePaths
+     * @return array
+     */
+    public static function scanControllers(string $directory, array $ignorePaths = []): array
+    {
+        // Set up directory iterators
+        $dirIterator = new RecursiveDirectoryIterator($directory);
+        $iterator = new RecursiveIteratorIterator($dirIterator);
 
-		// get all controllers
-		$allControllers = new \RegexIterator(
-			$it,
-			'/^.+Controller\.php$/',
-			\RecursiveRegexIterator::GET_MATCH
-		);
-		$controllers = array_keys(iterator_to_array($allControllers));
+        // Find all controller files
+        $allControllers = new RegexIterator(
+            $iterator,
+            '/^.+Controller\.php$/',
+            RecursiveRegexIterator::GET_MATCH
+        );
 
-		// apply ignore path
-		if (! empty($ignorePath)) {
-			if (is_array($ignorePath)) {
-				$ignoreRegexp = '('.implode('|', $ignorePath).')';
-			} else {
-				$ignoreRegexp = "({$ignorePath})";
-			}
+        $controllerPaths = array_keys(iterator_to_array($allControllers));
 
-			// find controllers, which should be ignored
-			$ignoredControllers = new \RegexIterator(
-				$allControllers,
-				"#^.+{$ignoreRegexp}.+Controller\.php$#",
-				\RecursiveRegexIterator::GET_MATCH,
-				\RecursiveRegexIterator::USE_KEY
-			);
+        // Filter out ignored paths
+        if (!empty($ignorePaths)) {
+            $ignorePattern = '(' . implode('|', $ignorePaths) . ')';
+            $ignoredControllers = new RegexIterator(
+                $allControllers,
+                "#^.+$ignorePattern.+Controller\.php$#",
+                RecursiveRegexIterator::GET_MATCH
+            );
 
-			$ignoredControllers = array_keys(iterator_to_array($ignoredControllers));
+            $ignoredPaths = array_keys(iterator_to_array($ignoredControllers));
+            $controllerPaths = array_diff($controllerPaths, $ignoredPaths);
+        }
 
-			$controllers = array_diff($allControllers, $ignoredControllers);
-		}
+        return $controllerPaths;
+    }
 
-		return $controllers;
-	}
+    /**
+     * Scans controller files to find their public action routes.
+     *
+     * @param array $controllerFiles
+     * @return array
+     */
+    public static function scanControllerActionIds(array $controllerFiles): array
+    {
+        $actions = [];
 
-	/**
-	 * Scan controllers by filenames to find their public action routes.
-	 *
-	 * @param array $controllers
-	 *
-	 * @return array
-	 */
-	public static function scanControllerActionIds(array $controllers)
-	{
-		$actions = [];
-		foreach ($controllers as $filename) {
-			$content = file_get_contents($filename);
-			$content = preg_replace(
-				"/(\/\*([^*]|[\r\n]|(\*+([^*\/]|[\r\n])))*\*\/)|(\/\/.*)/i",
-				'',
-				$content
-			);
-			// ignore abstract classes
-			if (false !== strpos($content, 'abstract class')) {
-				continue;
-			}
+        foreach ($controllerFiles as $filePath) {
+            $namespace = self::getNamespaceFromFile($filePath);
+            $className = self::getClassNameFromFile($filePath);
 
-			if (! preg_match('/namespace\s+([a-z0-9_\\\\]+)/i', $content, $namespaceMatch)) {
-				continue;
-			}
+            if ($namespace === null || $className === null) {
+                continue;
+            }
 
-			if (! preg_match('/class\s+(([a-z0-9_]+)Controller)[^{]+{/i', $content, $classMatch)) {
-				continue;
-			}
+            $fullyQualifiedClassName = "\\$namespace\\$className";
 
-			$className = '\\' . $namespaceMatch[1] . '\\' . $classMatch[1];
-			$reflection = new \ReflectionClass($className);
+            try {
+                $reflection = new ReflectionClass($fullyQualifiedClassName);
+            } catch (ReflectionException) {
+                Yii::warning("RBAC Scanner: Failed to reflect class $fullyQualifiedClassName.");
+                continue;
+            }
 
-			// ignore console commands
-			if (! $reflection->isSubclassOf(\yii\web\Controller::className())) {
-				continue;
-			}
+            if (!$reflection->isSubclassOf(Controller::class)) {
+                continue;
+            }
 
-			// find public methods
-			$methods = $reflection->getMethods(\ReflectionMethod::IS_PUBLIC);
+            $moduleId = self::getModuleIdFromNamespace($reflection->getNamespaceName());
+            $controllerId = Inflector::slug(Inflector::camel2words($className));
 
-			$moduleId = '';
-			if (preg_match('/modules\\\\([a-z0-9_-]+)\\\\/i', $reflection->getNamespaceName(), $moduleMatch)) {
-				$moduleId = Inflector::slug(Inflector::camel2words($moduleMatch[1])) . '/';
-			}
+            $actions = array_merge($actions, self::getPublicActions($reflection, $moduleId, $controllerId));
+        }
 
-			foreach ($methods as $method) {
-				if (! preg_match('/^action([A-Z]([a-zA-Z0-9]+))$/', $method->getName(), $actionMatch)
-				    && !('actions' === $method->getName() && $reflection->getName() === $method->class)
-				) {
-					continue;
-				}
-				$controllerId = Inflector::slug(Inflector::camel2words($classMatch[2]));
+        return $actions;
+    }
 
-				if ('actions' === $method->getName()) {
-					try {
-						$controllerObj = Yii::createObject($method->class, [$controllerId, Yii::$app]);
-						$customActions = $controllerObj->actions();
-						foreach ($customActions as $actionId => $params) {
-							$actions[] = $moduleId . $controllerId . '/' . $actionId;
-						}
-					} catch (\Exception $e) {
-						Yii::warning("RBAC Scanner: can't scan custom actions from {$method->class}::actions(). You will need to add them manually.");
-					}
+    /**
+     * Extracts the namespace from a file's content.
+     *
+     * @param string $filePath
+     * @return string|null
+     */
+    private static function getNamespaceFromFile(string $filePath): ?string
+    {
+        $content = file_get_contents($filePath);
+        if (preg_match('/namespace\s+([a-z0-9_\\\\]+)/i', $content, $match)) {
+            return $match[1];
+        }
+        return null;
+    }
 
-				} else {
-					$actionId = Inflector::slug(Inflector::camel2words($actionMatch[1]));
-					$actions[] = $moduleId . $controllerId . '/' . $actionId;
-				}
-			}
-		}
+    /**
+     * Extracts the class name from a file's content.
+     *
+     * @param string $filePath
+     * @return string|null
+     */
+    private static function getClassNameFromFile(string $filePath): ?string
+    {
+        $content = file_get_contents($filePath);
+        if (preg_match('/class\s+([a-z0-9_]+)Controller/i', $content, $match)) {
+            return $match[1];
+        }
+        return null;
+    }
 
-		return $actions;
-	}
+    /**
+     * Extracts the module ID from a namespace.
+     *
+     * @param string $namespace
+     * @return string
+     */
+    private static function getModuleIdFromNamespace(string $namespace): string
+    {
+        if (preg_match('/modules\\\\([a-z0-9_-]+)\\\\/i', $namespace, $match)) {
+            return Inflector::slug(Inflector::camel2words($match[1])) . '/';
+        }
+        return '';
+    }
+
+    /**
+     * Retrieves public action routes from a reflected controller.
+     *
+     * @param ReflectionClass $reflection
+     * @param string $moduleId
+     * @param string $controllerId
+     * @return array
+     */
+    private static function getPublicActions(ReflectionClass $reflection, string $moduleId, string $controllerId): array
+    {
+        $actions = [];
+        $methods = $reflection->getMethods(ReflectionMethod::IS_PUBLIC);
+
+        foreach ($methods as $method) {
+            $methodName = $method->getName();
+
+            if (preg_match('/^action([A-Z][a-zA-Z0-9]*)$/', $methodName, $match)) {
+                $actionId = Inflector::slug(Inflector::camel2words($match[1]));
+                $actions[] = $moduleId . $controllerId . '/' . $actionId;
+            } elseif ($methodName === 'actions' && $reflection->getName() === $method->class) {
+                $actions = array_merge($actions, self::getCustomActions($method, $controllerId, $moduleId));
+            }
+        }
+
+        return $actions;
+    }
+
+    /**
+     * Retrieves custom actions from the `actions()` method of a controller.
+     *
+     * @param ReflectionMethod $method
+     * @param string $controllerId
+     * @param string $moduleId
+     * @return array
+     */
+    private static function getCustomActions(ReflectionMethod $method, string $controllerId, string $moduleId): array
+    {
+        $actions = [];
+        try {
+            $controller = Yii::createObject($method->class, [$controllerId, Yii::$app]);
+            $customActions = $controller->actions();
+
+            foreach ($customActions as $actionId => $params) {
+                $actions[] = $moduleId . $controllerId . '/' . $actionId;
+            }
+        } catch (Exception) {
+            Yii::warning("RBAC Scanner: Failed to scan custom actions from $method->class::actions().");
+        }
+
+        return $actions;
+    }
 }
